RENAME to "mujoco-toolbox"
- refactor MjWrapper code to Wrapper.
- Relocate static controller code to Controller.py
